vaults move: ../data_sources/[dsId]/documents/index
