+++ front/pages/api/v1/w/[wId]/vaults/[vId]/data_sources/[dsId]/search.ts
@@ -182,7 +182,6 @@ type DatasourceSearchResponseBody = {
  *       405:
  *         description: Method not supported error
  */
-
 async function handler(
   req: NextApiRequest,
   res: NextApiResponse<WithAPIErrorResponse<DatasourceSearchResponseBody>>,
